+++ src/common_clj/component/helper/core.clj
@@ -1,0 +1,7 @@
+(ns common-clj.component.helper.core
+  (:require [schema.core :as s]))
+
+(s/defn get-component-content
+  [component :- s/Keyword
+   system]
+  (get-in system [component component]))